Action plugin for timestamping a file
This change adds an action plugin called timestamp_file that will
check if a file exists on a remote system and copies it out of the
way so that a new version can be created. This plugin can be used to
ensure that logs aren't overwritten or some other file you want to keep
before modifying.

Change-Id: I699641f65f20c7c4814632eb380525843ebcba98