Global role readme
We're getting to the part where we should have a general readme for the
roles that contains documentation about some global variables that might
be consumed across multiple roles by default. Currently this is limited
to the auth related variables.  Additionally the readme has a basic
example undercloud install playbook to indicate how some of the roles
can be used.

Change-Id: I511b6ace42a08e8669d727aa19c5bc4b51c069a5