Add new parameter in order to get a kind of dry-run
A new global parameter is introduced:
tripleo_generate_scripts_only

It allows to get the roles running as usual, but it won't actually
execute the "shell" module.
With the "tripleo_generate_scripts", it allows to get everything ready
for manual investigations, with the generated scripts ready.

Of course, every role can be instructed to only generate scripts, using
the standard <role_name>_generate_scripts_only parameter.

Change-Id: I841743fb6c2c24e4a3d86598f99322022b27804d