Added publish-openstack-python-tarball job to publish the tarballs
We are working on consuming tripleo-operator-ansible
as a rpm from stein onwards in deployment via tripleo-quickstart.

In order to build the package for different release, we need
tarball for tripleo-ansible-operator. Currently only tarballs
are getting created from master branch[1].

We need the tarballs from tagged release. In order to do that
we can add the publish-openstack-python-tarball
job[2] to publish the tarball.

[1]. https://tarballs.opendev.org/openstack/tripleo-operator-ansible/

[2]. https://review.opendev.org/c/openstack/project-config/+/791745

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I8b77429737c0256d75fe04fa6535e99a876c9d30