Add tripleo_upgrade role.
Includes the role to automate the execution of the openstack
tripleo upgrade command. As most of the code is shared with
tripleo-deploy this role is mostly based on tripleo-deploy.
The only appreciable difference is that the --upgrade parameter
is not being passed (as that is being passed to true when
invoking the Deploy class [0]).

[0] - https://github.com/openstack/python-tripleoclient/blob/master/tripleoclient/v1/tripleo_upgrade.py#L41

Change-Id: Ie2eb2c220fef9284bf5fa6ef6908a00839589649